+++ app/src/marketplace/containers/EditProductPage/SaveProductDialog/index.jsx
@@ -9,12 +9,13 @@ import { selectTransactionState as selectUpdateTransactionState } from '$mp/modu
 import { selectUpdateProductTransaction, selectUpdateContractProductError } from '$mp/modules/updateContractProduct/selectors'
 import { saveProduct, resetSaveDialog } from '$mp/modules/saveProductDialog/actions'
 import { hideModal } from '$mp/modules/modals/actions'
-import { transactionStates, saveProductSteps } from '$mp/utils/constants'
+import { saveProductSteps } from '$mp/utils/constants'
+import { transactionStates } from '$shared/utils/constants'
 import type { StoreState, SaveProductStep } from '$mp/flowtype/store-state'
 import type { ProductId } from '$mp/flowtype/product-types'
 import type { TransactionState } from '$mp/flowtype/common-types'
 import type { ErrorInUi } from '$shared/flowtype/common-types'
-import type { TransactionEntity } from '$mp/flowtype/web3-types'
+import type { TransactionEntity } from '$shared/flowtype/web3-types'
 import withContractProduct from '$mp/containers/WithContractProduct'
 import { selectStep, selectUpdateFinished } from '$mp/modules/saveProductDialog/selectors'
