+++ app/src/marketplace/containers/ProductPage/PublishOrUnpublishDialog/PublishDialog/index.jsx
@@ -9,7 +9,8 @@ import CompletePublishDialog from '$mp/components/Modal/CompletePublishDialog'
 import CompleteContractProductPublishDialog from '$mp/components/Modal/CompleteContractProductPublishDialog'
 import NoStreamsWarningDialog from '$mp/components/Modal/NoStreamsWarningDialog'
 import { formatPath } from '$shared/utils/url'
-import { publishFlowSteps, transactionStates } from '$mp/utils/constants'
+import { publishFlowSteps } from '$mp/utils/constants'
+import { transactionStates } from '$shared/utils/constants'
 import { selectStep } from '$mp/modules/publishDialog/selectors'
 import { publishOrCreateProduct } from '$mp/modules/publishDialog/actions'
 import { selectFetchingContractProduct } from '$mp/modules/contractProduct/selectors'
@@ -24,7 +25,7 @@ import { selectFetchingProduct } from '$mp/modules/product/selectors'
 import type { StoreState, PublishStep } from '$mp/flowtype/store-state'
 import type { TransactionState } from '$mp/flowtype/common-types'
 import type { Product, ProductId } from '$mp/flowtype/product-types'
-import type { TransactionEntity } from '$mp/flowtype/web3-types'
+import type { TransactionEntity } from '$shared/flowtype/web3-types'
 import type { ErrorInUi } from '$shared/flowtype/common-types'
 
 type StateProps = {
