+++ app/src/marketplace/flowtype/common-types.js
@@ -1,15 +1,12 @@
 // @flow
 
-import { currencies, timeUnits, transactionStates, transactionTypes, notificationIcons } from '../utils/constants'
-import type { Hash } from '../flowtype/web3-types'
+import { currencies, timeUnits, notificationIcons } from '../utils/constants'
+
+import type { Hash } from '$shared/flowtype/web3-types'
 
 export type Currency = $Values<typeof currencies>
 
 export type TimeUnit = $Values<typeof timeUnits>
-
-export type TransactionState = $Values<typeof transactionStates>
-
-export type TransactionType = $Values<typeof transactionTypes>
 
 export type NumberString = string // Must be parsable to BigNumber
