+++ app/src/marketplace/flowtype/store-state.js
@@ -2,7 +2,7 @@
 
 import { purchaseFlowSteps, publishFlowSteps, saveProductSteps } from '../utils/constants'
 
-import TransactionError from '../errors/TransactionError'
+import TransactionError from '$shared/errors/TransactionError'
 import type { CategoryIdList } from './category-types'
 import type {
     EditProduct,
@@ -12,11 +12,10 @@ import type {
     Subscription,
     ProductPermissions,
 } from './product-types'
-import type { Hash, Address, HashList } from './web3-types'
+import type { Hash, Address, HashList } from '$shared/flowtype/web3-types'
 import type { StreamIdList } from '$shared/flowtype/stream-types'
-import type { Purchase, TransactionState, Notification, NumberString } from './common-types'
-import type { ErrorInUi } from '$shared/flowtype/common-types'
-import type { EntitiesState, UserState } from '$shared/flowtype/store-state'
+import type { Purchase, Notification, NumberString } from './common-types'
+import type { ErrorInUi, TransactionState } from '$shared/flowtype/common-types'
 
 // categories
 export type CategoryState = {
@@ -227,7 +226,6 @@ export type StoreState = {
     contractProduct: ContractProductState,
     createContractProduct: ModifyContractProductState,
     editProduct: EditProductState,
-    entities: EntitiesState,
     global: GlobalState,
     i18n: I18nState,
     modals: ModalState,
@@ -245,7 +243,6 @@ export type StoreState = {
     relatedProducts: RelatedProductListState,
     streams: StreamsState,
     updateContractProduct: ModifyContractProductState,
-    user: UserState,
     web3: Web3State,
     transactions: TransactionsState,
 }
