+++ app/src/marketplace/modules/allowance/selectors.js
@@ -7,7 +7,7 @@ import type { AllowanceState, StoreState } from '../../flowtype/store-state'
 import type { EntitiesState } from '$shared/flowtype/store-state'
 import type { NumberString } from '../../flowtype/common-types'
 import type { ErrorInUi } from '$shared/flowtype/common-types'
-import type { Hash, TransactionEntity } from '../../flowtype/web3-types'
+import type { Hash, TransactionEntity } from '$shared/flowtype/web3-types'
 import { transactionSchema } from '$shared/modules/entities/schema'
 import { selectEntities } from '$shared/modules/entities/selectors'
