+++ app/src/marketplace/modules/allowance/types.js
@@ -2,7 +2,7 @@
 
 import type { NumberString } from '../../flowtype/common-types'
 import type { PayloadAction, ErrorInUi } from '$shared/flowtype/common-types'
-import type { Hash } from '../../flowtype/web3-types'
+import type { Hash } from '$shared/flowtype/web3-types'
 
 export type AllowanceAction = PayloadAction<{
     allowance: NumberString,
