+++ app/src/marketplace/modules/createContractProduct/actions.js
@@ -5,12 +5,12 @@ import { getLocation } from 'react-router-redux'
 
 import { setProductDeploying } from '$mp/modules/publish/actions'
 import { getProductById } from '$mp/modules/product/actions'
-import type { Hash } from '$mp/flowtype/web3-types'
+import type { Hash } from '$shared/flowtype/web3-types'
 import type { ProductId, SmartContractProduct } from '$mp/flowtype/product-types'
 import type { ErrorInUi, ReduxActionCreator } from '$shared/flowtype/common-types'
 import type { StoreState } from '$mp/flowtype/store-state'
 import { addTransaction } from '$mp/modules/transactions/actions'
-import { transactionTypes } from '$mp/utils/constants'
+import { transactionTypes } from '$shared/utils/constants'
 
 import * as services from './services'
 import {
