+++ app/src/marketplace/modules/createContractProduct/selectors.js
@@ -6,7 +6,7 @@ import { denormalize } from 'normalizr'
 
 import type { ModifyContractProductState, StoreState } from '$mp/flowtype/store-state'
 import type { EntitiesState } from '$shared/flowtype/store-state'
-import type { Hash, TransactionEntity } from '$mp/flowtype/web3-types'
+import type { Hash, TransactionEntity } from '$shared/flowtype/web3-types'
 import { transactionSchema } from '$shared/modules/entities/schema'
 import { selectEntities } from '$shared/modules/entities/selectors'
 import type { ErrorInUi } from '$shared/flowtype/common-types'
