+++ app/src/marketplace/utils/constants.js
@@ -66,23 +66,6 @@ export const timeUnits = {
     month: 'month',
 }
 
-export const transactionStates = {
-    STARTED: 'started', // transaction started
-    PENDING: 'pending', // hash received
-    CONFIRMED: 'confirmed', // mined
-    FAILED: 'failed', // error
-}
-
-export const transactionTypes = {
-    SET_ALLOWANCE: 'setAllowance',
-    RESET_ALLOWANCE: 'resetAllowance',
-    PURCHASE: 'purchase',
-    CREATE_CONTRACT_PRODUCT: 'createContractProduct',
-    UPDATE_CONTRACT_PRODUCT: 'updateContractProduct',
-    REDEPLOY_PRODUCT: 'redeployProduct',
-    UNDEPLOY_PRODUCT: 'undeployProduct',
-}
-
 export const productListPageSize = 20
 
 export const gasLimits = {
