+++ app/src/shared/errors/TransactionError/index.js
