+++ app/src/shared/flowtype/common-types.js
@@ -1,5 +1,7 @@
 // @flow
 
+import { transactionStates, transactionTypes } from '$shared/utils/constants'
+
 export type ErrorInUi = {
     message: string,
     statusCode?: ?number,
@@ -21,3 +23,7 @@ export type RequestMethod = 'get' | 'post' | 'put' | 'delete'
 export type PropertySetter<T> = (string, T) => void
 
 export type ApiResult<T> = Promise<T>
+
+export type TransactionState = $Values<typeof transactionStates>
+
+export type TransactionType = $Values<typeof transactionTypes>
