+++ app/src/shared/modules/entities/selectors.js
@@ -1,6 +1,5 @@
 // @flow
 
-import type { EntitiesState } from '$shared/flowtype/store-state'
-import type { StoreState } from '$mp/flowtype/store-state'
+import type { EntitiesState, StoreState } from '$shared/flowtype/store-state'
 
 export const selectEntities = (state: StoreState): EntitiesState => state.entities
