+++ app/src/shared/modules/integrationKey/constants.js
@@ -1,0 +1,5 @@
+// @flow
+
+export const INTEGRATION_KEYS_REQUEST: string = 'marketplace/integrationKey/INTEGRATION_KEYS_REQUEST'
+export const INTEGRATION_KEYS_SUCCESS: string = 'marketplace/integrationKey/INTEGRATION_KEYS_SUCCESS'
+export const INTEGRATION_KEYS_FAILURE: string = 'marketplace/integrationKey/INTEGRATION_KEYS_FAILURE'
