+++ app/src/shared/modules/user/services.js
@@ -5,14 +5,12 @@ import zxcvbn from 'zxcvbn'
 import { get, post } from '$shared/utils/api'
 import { formatApiUrl } from '$shared/utils/url'
 import type { ApiResult } from '$shared/flowtype/common-types'
-import type { User, IntegrationKey, ApiKey, PasswordUpdate } from '$shared/flowtype/user-types'
+import type { User, ApiKey, PasswordUpdate } from '$shared/flowtype/user-types'
 // import routes from '$routes'
 
 export const getMyKeys = (): ApiResult<Array<ApiKey>> => get(formatApiUrl('users', 'me', 'keys', {
     noCache: Date.now(),
 }))
-
-export const getIntegrationKeys = (): ApiResult<Array<IntegrationKey>> => get(formatApiUrl('integration_keys'))
 
 export const getUserData = (): ApiResult<User> => get(formatApiUrl('users', 'me', {
     noCache: Date.now(),
