+++ app/src/shared/web3/abis/token.json
