+++ app/src/shared/web3/token.config.js
